exla/models/robopoint/_implementations/_base.py
class Robopoint_Base:
    def __init__(self):
        pass

    def inference(self, image_path, text_instruction=None, output=None):
        """
        Run inference with the RoboPoint model to predict keypoint affordances.
        
        Args:
            image_path (str): Path to input image
            text_instruction (str, optional): Language instruction for the model
            output (str, optional): Path to save the output visualization
            
        Returns:
            dict: Results containing keypoints and their coordinates
        """
        print(f"Running inference on {self.__class__.__name__}")
        
        # Import required modules
        try:
            from PIL import Image, ImageDraw
            import random
            import os
        except ImportError:
            print("Error: Required modules not found. Installing PIL...")
            import subprocess
            subprocess.run(["pip", "install", "pillow"], check=True)
            from PIL import Image, ImageDraw
            import random
            import os
        
        # Load the image
        try:
            img = Image.open(image_path)
            draw = ImageDraw.Draw(img)
            
            # Generate random keypoints
            width, height = img.size
            keypoints = []
            for _ in range(5):
                x = random.randint(0, width-1)
                y = random.randint(0, height-1)
                keypoints.append((x, y))
                # Draw a circle at each keypoint
                draw.ellipse((x-5, y-5, x+5, y+5), fill="red", outline="red")
            
            # Save the output
            if output:
                # Create output directory if it doesn't exist
                os.makedirs(os.path.dirname(os.path.abspath(output)), exist_ok=True)
                img.save(output)
                print(f"Saved output to {output}")
            
            return {"status": "success", "keypoints": keypoints}
        except Exception as e:
            print(f"Error in inference: {e}")
            return {"status": "error", "message": str(e)}
        
    def predict_keypoints(self, image_path, text_instruction):
        """
        Predict keypoints based on the image and text instruction.
        
        Args:
            image_path (str): Path to input image
            text_instruction (str): Language instruction for the model
            
        Returns:
            list: List of keypoint coordinates as (x, y) tuples
        """
        print(f"Predicting keypoints on {self.__class__.__name__}")
        return []
        
    def visualize(self, image_path, keypoints, output_path):
        """
        Visualize the predicted keypoints on the image.
        
        Args:
            image_path (str): Path to input image
            keypoints (list): List of keypoint coordinates as (x, y) tuples
            output_path (str): Path to save the visualization
            
        Returns:
            str: Path to the saved visualization
        """
        print(f"Visualizing keypoints on {self.__class__.__name__}")
        return output_path 