exla/models/robopoint/_implementations/requirements/requirements_gpu.txt
numpy>=1.24.0
pillow>=9.0.0
transformers>=4.31.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
safetensors>=0.3.1
protobuf>=3.20.0
einops>=0.6.1
timm>=0.9.5